Refactor out DistributedTargets and clean up.
